#D8-130
Go through the project and ensure all file names match the naming-convention and all hyperlinks are linked to a page correctly.
