+++ components/sign-in-sign-out-bar.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import styles from '../styles/components_Styles/SignInSignOutBarStyles.module.css'
+import styles from '../styles/components_Styles/sign-in-sign-out-bar-styles.module.css'
 import Link from 'next/link'
 import { useSession } from "next-auth/react"
 import { signOut } from "next-auth/react"
@@ -15,8 +15,8 @@ export default function SignInSignOutBar() {
     if (status === "loading") {
       return <nav className={styles.Container_Main}>
       <ul className={styles.SignInSignOut_ul}>
-          <Link className={styles.SignInSignOut_li} href ='../signInPage'><li>Sign In</li></Link>
-          <Link className={styles.SignInSignOut_li} href ='../createAccountPage'><li>Create Account</li></Link> 
+          <Link className={styles.SignInSignOut_li} href ='../sign-in-page'><li>Sign In</li></Link>
+          <Link className={styles.SignInSignOut_li} href ='../create-account-page'><li>Create Account</li></Link> 
       </ul>
