+++ pages/engraving-design.js
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import styles from '../styles/hennashow.module.css';
+import styles from '../styles/henna-show.module.css';
 
 function EngravingGallery({designs}){
     return (
