+++ pages/index.js
@@ -2,7 +2,7 @@ import Head from "next/head";
 import Image from "next/image";
 import Link from 'next/link';
 import { Inter } from "next/font/google";
-import styles from "@/styles/Homepage.module.css";
+import styles from "@/styles/homepage.module.css";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -34,7 +34,7 @@ export default function Home() {
                         
                     </section>
                     <section className={styles.gridColor3}> 
-                        <Link href='./engraving-Showcase'><h1 className={styles.gridText}>Engraving</h1></Link>
+                        <Link href='./engraving-design'><h1 className={styles.gridText}>Engraving</h1></Link>
                         
                     </section>
